refactor(presets): simplify create presets codes
