+++ e2e/__tests__/ts-jest-presets.test.ts
@@ -1,0 +1,13 @@
+import { allValidPackageSets } from '../__helpers__/templates'
+import { configureTestCase } from '../__helpers__/test-case'
+
+describe('ts-jest presets', () => {
+  const testCase = configureTestCase('presets')
+
+  testCase.runWithTemplates(allValidPackageSets, 0, (runTest, { testLabel }) => {
+    it(testLabel, () => {
+      const result = runTest()
+      expect(result.status).toBe(0)
+    })
+  })
+})
