--- presets/create.js
@@ -1,1 +1,0 @@
-module.exports = require('../dist/config/create-jest-preset').createJestPreset
