+++ presets/index.js
@@ -1,20 +1,1 @@
-const create = require('./create')
-
-module.exports = {
-  get defaults() {
-    return create()
-  },
-  get jsWithTs() {
-    return create({ allowJs: true })
-  },
-  get jsWithBabel() {
-    return create(
-      { allowJs: false },
-      {
-        transform: {
-          '^.+\\.jsx?$': 'babel-jest',
-        },
-      }
-    )
-  },
-}
+module.exports = require('../dist/presets')
